+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Compatibility:
 * Add `rb_category_warn` function (#3710, @andrykonchin).
 * Add `rb_gc_mark_locations()` (#3704, @andrykonchin).
 * Implement `rb_str_format()` (#3716, @andrykonchin).
+* Add `IO#{pread, pwrite}` methods (#3718, @andrykonchin).
 
 Performance: